refactor: swagger interface 분리
