+++ src/main/java/com/sprint/mission/findex/domain/syncjob/dto/SyncJobResponse.java
